add histogram endpoint and security test

+++ server/src/main/java/com/seebie/server/dto/HistogramNormalized.java
@@ -2,6 +2,6 @@ package com.seebie.server.dto;
 
 import java.util.List;
 
-public record HistogramNormalized(List<Integer> bins, List<List<Double>> dataSets) {
+public record HistogramNormalized(List<Integer> bins, List<List<Integer>> dataSets) {
 
 }
